rename: store 파일 접두사에 use를 붙여서 컨벤션 통일
